requirements.txt
aniso8601==9.0.1
asgiref==3.8.1
beautifulsoup4==4.12.3
blinker==1.8.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
defusedxml==0.8.0rc2
Django==5.0.4
django-allauth==0.63.0
django-extensions==3.2.3
django-htmx==1.17.3
django-phonenumber-field==7.3.0
djangorestframework==3.15.1
Flask==3.0.3
Flask-RESTful==0.3.10
gunicorn==22.0.0
idna==3.7
importlib_metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
oauthlib==3.2.2
packaging==24.0
phonenumbers==8.13.36
pillow==10.3.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
python-dotenv==1.0.1
python-stdnum==1.20
python3-openid==3.2.0
pytz==2024.1
requests==2.31.0
requests-oauthlib==2.0.0
six==1.16.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.2
zipp==3.18.1

